refactor: moved AppConfig to config directory
