+++ sites/geohub/src/lib/config/AppConfig/AcceptedExtensions.ts
