+++ sites/geohub/src/lib/config/AppConfig/DataCategories.ts
