+++ sites/geohub/src/lib/config/AppConfig/FooterItems.ts
