+++ sites/geohub/src/lib/config/AppConfig/HeaderItems.ts
