+++ sites/geohub/src/lib/config/AppConfig/MapSortingColumns.ts
