+++ sites/geohub/src/lib/config/AppConfig/MapStyles.ts
