+++ sites/geohub/src/lib/config/AppConfig/TagInputValues.ts
