+++ sites/geohub/src/lib/config/AppConfig/attribution.ts
