+++ sites/geohub/src/routes/data/publish/+page.server.ts
@@ -11,7 +11,7 @@ import {
 } from '$lib/server/helpers'
 import { removeSasTokenFromDatasetUrl } from '$lib/helper'
 import { env } from '$env/dynamic/private'
-import { Permission } from '$lib/AppConfig'
+import { Permission } from '$lib/config/AppConfig'
 
 /**
  * Preload dataset metadata from either database (existing case) or titiler/pmtiles (new case)
